requirements.txt
streamlit
Pillow
google-cloud-vision
google-cloud-firestore
rapidfuzz